% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adirondack}
\alias{adirondack}
\title{Adjacency matrix of the Adirondack lakes}
\format{
A matrix with 211 rows and columns with 1 if species interacts and 0
  if they don't. The names of the species are saved as matrix row and
  column-names.
}
\source{
\url{https://idata.idiv.de/ddm/Data/ShowData/283?version=3}
}
\usage{
adirondack
}
\description{
A dataset containing the binary predator-prey interactions for the Adirondack
lakes, obtained from the GATEWAy database 1.0 (Brose).
}
\keyword{datasets}
