% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected.R
\name{connected_pool}
\alias{connected_pool}
\title{Create random pool of connected species}
\usage{
connected_pool(web, S, verbose = FALSE, max_iter = 1000, plot = TRUE, iter = 0)
}
\arguments{
\item{web}{adjacency matrix.}

\item{S}{number of species.}

\item{verbose}{TRUE/FALSE.}

\item{max_iter}{maximum number of iterations.}

\item{plot}{TRUE/FALSE if to plot the community.}
}
\description{
Create random pool of connected species
}
\examples{
a <- connected_pool(web, 13, verbose = FALSE)
g <- graph_from_adjacency_matrix(web[a, a], "undirected")
while (igraph::components(g, "strong")$no > 1) {
  message("Not strongly connected - rerunning")
  a <- connected_pool(web, 13, verbose = FALSE)
  g <- graph_from_adjacency_matrix(web[a, a], "undirected")
}
plot(g)
g <- graph_from_adjacency_matrix(web[a, a], "directed")
all(as.numeric(names(as.numeric(names((degree(g, mode = "in") == 0))))) \%in\% basal)
}
