% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.basals}
\alias{.basals}
\title{Basal species in the metaweb}
\usage{
.basals(metaweb)
}
\arguments{
\item{metaweb}{adjacency matrix of the meta food web (metaweb).}
}
\value{
vectors of basal species names.
}
\description{
Basal species in the metaweb
}
