% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.components}
\alias{.components}
\title{Number of connected components}
\usage{
.components(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector of species names.}

\item{metaweb}{adjacency matrix of the metaweb.}
}
\description{
Number of connected components
}
\details{
Return the number of connected components in the community
}
