% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.consumers}
\alias{.consumers}
\title{Consumers in the metaweb}
\usage{
.consumers(metaweb)
}
\arguments{
\item{metaweb}{adjacency matrix of the meta food web (metaweb).}
}
\value{
vectors of basal species names.
}
\description{
Consumers in the metaweb
}
