% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.find_isolated}
\alias{.find_isolated}
\title{Find isolated species}
\usage{
.find_isolated(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector of species names.}

\item{metaweb}{adjacency matrix of the metaweb.}
}
\value{
a vector with the names of isolated basal species and isolated consumers.
}
\description{
Find isolated species
}
