% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.find_replacements}
\alias{.find_replacements}
\title{Find replacement species}
\usage{
.find_replacements(sp.names, isolated, metaweb, keep.n.basal = FALSE)
}
\arguments{
\item{sp.names}{vector of species names.}

\item{isolated}{vector of species names that are isolated.}

\item{metaweb}{adjacency matrix of the metaweb.}

\item{keep.n.basal}{logical, if to keep the constant number of basal species.}
}
\value{
a vector or replacement names.
}
\description{
Find replacement species
}
