% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_adj}
\alias{.get_adj}
\title{Get food web adjacency matrix}
\usage{
.get_adj(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}
}
\value{
The adjacency matrix of the local community
}
\description{
Get food web adjacency matrix
}
