% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.modules}
\alias{.modules}
\title{Number of modules of the food web}
\usage{
.modules(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector of species names.}

\item{metaweb}{adjacency matrix of the metaweb.}
}
\description{
Number of modules of the food web
}
\details{
Return the number of modules in the network. Modules are obtained
  using a fast greedy algorithm on the undirected food web.
}
