% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.top}
\alias{.top}
\title{Top consumer species in the metaweb}
\usage{
.top(metaweb)
}
\arguments{
\item{metaweb}{adjacency matrix of the meta food web (metaweb).}
}
\value{
vectors of top consumer species names.
}
\description{
Top consumer species in the metaweb
}
