% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_draws.R
\name{draw_random_species}
\alias{draw_random_species}
\title{Draw random species from a meta-community}
\usage{
draw_random_species(n, sp.names = NULL, S = NULL)
}
\arguments{
\item{n}{integer, number of species to draw.}

\item{sp.names}{vector, names of the species in the meta-community (default =
NULL).}

\item{S}{integer, meta-community species richness (default = NULL).}
}
\value{
A vector with the species names.
}
\description{
Draw random species from a meta-community
}
\details{
Either one of \code{sp.names} or \code{S} need to be specified. If
  both are specified, \code{sp.names} overrules \code{S}, which will not be
  evaluated.
}
\examples{
S <- 10
species <- LETTERS[seq_len(S)]
draw_random_species(3, sp.names = species)
draw_random_species(3, S = S)
}
