% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_similarity_filtering.R
\name{metropolis.hastings}
\alias{metropolis.hastings}
\title{Metropolis-Hastings algorithm}
\usage{
metropolis.hastings(old.value, new.value, t = 0)
}
\arguments{
\item{old.value}{is the similarity value of the starting species pool.}

\item{new.value}{is the similarity value of the changed species pool.}

\item{t}{is the 'temperature' of the system.}
}
\value{
boolean, if to accept the replacement or not.
}
\description{
Metropolis-Hastings algorithm
}
