% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_metaweb}
\alias{name_metaweb}
\title{Name metaweb species}
\usage{
name_metaweb(metaweb)
}
\arguments{
\item{metaweb}{adjacency matrix of the metaweb, to use to order basal species first.}
}
\value{
the named metaweb.
}
\description{
Name metaweb species
}
\details{
Gives the metaweb matrix rownames and columnames drawing randomly
  from a sequene of letters.
}
