% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metrics.R
\name{network}
\alias{network}
\title{Compute network metrics}
\usage{
network(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}
}
\value{
a data.frame with network metrics.
}
\description{
Compute network metrics
}
