% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected.R
\name{replace_species}
\alias{replace_species}
\title{Replace species}
\usage{
replace_species(species_list, replacements)
}
\arguments{
\item{species_list}{vector of species with some NAs to be replaced.}

\item{replacements}{candidates for replacement.}
}
\value{
the original vector \code{species_list} with replaced species.
}
\description{
Replace species
}
