% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_filtering.R
\name{resource_filtering}
\alias{resource_filtering}
\title{Resource filtering assembly}
\usage{
resource_filtering(sp.names, metaweb, keep.n.basal = FALSE)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}

\item{keep.n.basal}{TRUE/FALSE, if to keep the number of basal species as
originally passed in sp.names.}
}
\value{
A vector with the species names.
}
\description{
Resource filtering assembly
}
