% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_fw}
\alias{show_fw}
\title{Plot food web}
\usage{
show_fw(sp.names, metaweb, title = NULL)
}
\arguments{
\item{sp.names}{adjancency matrix of the food web.}

\item{metaweb}{adjacency matrix of the metaweb, to use to order basal species first.}

\item{title}{string, title of the plot.}
}
\description{
Plot food web
}
