% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_similarity_filtering.R
\name{similarity_filtering}
\alias{similarity_filtering}
\title{Limiting similarity filtering assembly}
\usage{
similarity_filtering(
  sp.names,
  metaweb,
  t = 0,
  method = "jaccard",
  stat = "mean",
  mode = "all",
  max.iter = 1000
)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}

\item{t}{is the 'temperature' of the system.}

\item{method}{character, same as in similarity (igraph). Options are
'jaccard', 'dice', and 'invlogweighted'.}

\item{stat}{character, statistic used to summarize similarity. Currently
available are c("mean", "sum", "max").}

\item{mode}{character, which edges are used to compute similarity. Currently
available are c("all", "in", "out").}

\item{max.iter}{is the maximum number of moves computed.}
}
\value{
A vector with the species names.
}
\description{
Limiting similarity filtering assembly
}
\details{
Similarities are calculated using igraph as matrices. To summarize
  these into species-level metrics, the argument "stat" is needed. When stat
  = "mean", probability of removal of species is proportional to the average
  of their similarities, etc. Global similarities, i.e. of the whole food web,
  are also summarized depending on the "stat" argument in a similar way.

  When mode = 'in', similarity is computed using only 'in' links, i.e. species 
  are considered similar if they share similar resources, but not necessarily 
  have similar consumers. The opposite is treu when mode = 'out'. When 
  mode = 'all' (default) all edges are considered.

  The temperature parameter 't' specifies the degree of stochasticity of the
  algorithm. For t > 0, an unfavourable move can be accepted, if it passes a
  probabilistic acceptance criterion. For t = 0, stochasticity is removed and
  the algorithm becomes purely deterministic. This may be result in some
  unwarranted behavior, e.g. strongly modular food webs.
}
