% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metrics.R
\name{trophic_levels}
\alias{trophic_levels}
\title{Calculate trophic level}
\usage{
trophic_levels(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}
}
\value{
numeric vector of trophic levels.
}
\description{
Calculate trophic level
}
