% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trophic_niche}
\alias{trophic_niche}
\title{Trophic niches}
\usage{
trophic_niche(sp.names, metaweb)
}
\arguments{
\item{sp.names}{vector of species names.}

\item{metaweb}{adjacency matrix of the metaweb.}
}
\value{
matrix with mean and variance of the trophic level of species'
  resources. These are both zero if the species is a basal resource.
}
\description{
Trophic niches
}
\details{
Calculate the trophic niche of species as the mean and variance of
  the trophic level of their resources.
}
