% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wJaccard}
\alias{wJaccard}
\title{Compute weighted Jaccard similarity}
\usage{
wJaccard(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
The matrix with Jaccard similarity
}
\description{
Compute weighted Jaccard similarity
}
