% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_similarity_filtering.R
\name{.move}
\alias{.move}
\title{One move of the limiting similarity algorithm}
\usage{
.move(
  sp.names,
  metaweb,
  t = 0,
  method = "jaccard",
  stat = "mean",
  mode = "all",
  return.similarity = FALSE
)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}

\item{t}{is the 'temperature' of the system.}

\item{method}{character, same as in similarity (igraph). Options are
'jaccard', 'dice', and 'invlogweighted'.}

\item{stat}{character, statistic used to summarize similarity. Currently
available are c("mean", "sum", "max").}

\item{mode}{character, which edges are used to compute similarity. Currently
available are c("all", "in", "out").}

\item{return.similarity}{logical, if to return the difference in similarity
score.}
}
\value{
A vector with the species names.
}
\description{
One move of the limiting similarity algorithm
}
\details{
Similarities are calculated using igraph as matrices. To summarize
  these into species-level metrics, the argument "stat" is needed. When stat
  = "mean", probability of removal of species is proportional to the average
  of their similarities, etc. Global similarities, i.e. of the whole food
  web, are also summarized depending on the "stat" argument in a similar way.

  When mode = 'in', similarity is computed using only 'in' links, i.e.
  species are considered similar if they share similar resources, but not
  necessarily have similar consumers. The opposite is treu when mode = 'out'.
  When mode = 'all' (default) all edges are considered.
}
