% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_similarity_filtering.R
\name{.move}
\alias{.move}
\title{One move of the limiting similarity algorithm}
\usage{
.move(sp.names, metaweb, t = 0, method = "jaccard")
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}

\item{t}{is the 'temperature' of the system.}

\item{method}{character, same as in similarity (igraph). Options are
'jaccard', 'dice', and 'invlogweighted'.}
}
\value{
A vector with the species names.
}
\description{
One move of the limiting similarity algorithm
}
