% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_similarity_filtering.R
\name{similarity_filtering}
\alias{similarity_filtering}
\title{Limiting similarity filtering assembly}
\usage{
similarity_filtering(sp.names, metaweb, t = 0, max.iter = 1000)
}
\arguments{
\item{sp.names}{vector, names of the species in the meta-community.}

\item{metaweb}{adjacency matrix of the meta food web (metaweb).}

\item{t}{is the 'temperature' of the system.}

\item{max.iter}{is the number of maximum iterations.}
}
\value{
A vector with the species names.
}
\description{
Limiting similarity filtering assembly
}
